psql_server.py:  Properly handle a connection closing
psql does two connections.
The first one is before the password prompt.
As soon as it receives AuthenticationCleartextPassword, it closes the connection and prompts the user for a password.
It then creates a new connection with the password.

The windows version of psql waits for the server to close the connection, and just sends zero data.
However, the Linux version actually closes the connection.